requirements.txt
python-telegram-bot==21.6
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pandas==2.2.2
numpy==1.26.4
pydantic==2.9.2
pytz==2024.1
websockets==12.0
